Improved styling of the week_fixtures table.
Fixed the width of the input number fields because the size attribute of
the field no longer works in HTML 5.

Fixed the width of the teamname column because the team name is never
longer than 32 characters.

Allowed the size of the performances input to float within the column.

Set the table width to 80%.