000043 - Configuration Module Too Complex
Added missing valid paths to list in get_path().

The individual parts of a compound path do not have to exist. Therefore
when get_path is called recursively -allow_not_exists is set to true.
this means the check only applies to the whole compound path.

Fixed a typo in Configuration.t.