000043 - Configuration Module Too Complex
Modified Configuration.pm so the it can hold the csv file for the
current request. This allows it to return the full finename for
the csv file.

Modified results_system.pl to pass the division parameter to Starter.pm
which uses it to set the csv file name in Configuration.pm.

Added a few simple tests.